frontend: move socket.ts to subdir
